Move evaluateGuess function + test to respective files
