+++ src/card.js
@@ -8,16 +8,6 @@ function createCard(id, question, answers, correctAnswer){
     }
 
 }
-
-function evaluateGuess(guess, correctAnswer) {
-    if(correctAnswer === guess){
-        return 'Correct!'
-    } else {
-        return 'Incorrect!'
-    }
-}
-
 module.exports = {
-    createCard,
-    evaluateGuess
+    createCard
 }
