+++ src/turn.js
@@ -1,0 +1,11 @@
+function evaluateGuess(guess, correctAnswer) {
+    if(correctAnswer === guess){
+        return 'Correct!'
+    } else {
+        return 'Incorrect!'
+    }
+}
+
+module.exports = {
+    evaluateGuess,
+}
