+++ test/Card-test.js
@@ -18,18 +18,4 @@ describe('card', function() {
     expect(card.correctAnswer).to.equal('object');
   });  
 
-  it('should state if answer is correct', function() {
-    let correctGuess = evaluateGuess('object', card.correctAnswer);
-    let expectedResponse = 'Correct!'
-
-    expect(correctGuess).to.equal(expectedResponse)
-
-  });
-
-  it('should state if answer is incorrect', function() {
-    let incorrectGuess = evaluateGuess('not object', card.correctAnswer);
-    let expectedResponse = 'Incorrect!'
-
-    expect(incorrectGuess).to.equal(expectedResponse)
-  });
 });
